# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  CreateKnowledgeDTO:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      description:
        type: string
      driver:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - configuration
    - driver
    - name
    type: object
  CreateKnowledgeIndexDTO:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      driver:
        maxLength: 100
        type: string
      name:
        maxLength: 255
        type: string
    required:
    - configuration
    - driver
    - name
    type: object
  CreateKnowledgeResponseDTO:
    properties:
      knowledge:
        $ref: '#/definitions/KnowledgeResponse'
    type: object
...
  DriversDTO:
    properties:
      source_drivers:
        items:
          $ref: '#/definitions/Info'
        type: array
    type: object
  Error:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  IndexStatus:
    enum:
    - pending
    - syncing
    - indexed
    - skipped
    - error
    type: string
    x-enum-varnames:
    - IndexedStatusPending
    - IndexedStatusSyncing
    - IndexedStatusIndexed
    - IndexedStatusSkipped
    - IndexedStatusError
  IndexedKnowledgeItem:
    properties:
      hash:
        type: string
      id:
        type: string
      index_id:
        type: string
      indexed_at:
        type: string
      knowledge_item_id:
        type: string
      last_error:
        type: string
      metadata:
        $ref: '#/definitions/JSONMap'
      reason:
        type: string
      status:
        $ref: '#/definitions/IndexStatus'
    type: object
  IndexedKnowledgeItemsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/IndexedKnowledgeItem'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
  Info:
    properties:
      configuration_schema:
        $ref: '#/definitions/JSONMap'
      description:
        type: string
      icon:
        type: string
      key:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - description
    - name
    type: object
  JSONMap:
    additionalProperties: {}
    type: object
  Knowledge:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      description:
        type: string
      driver:
        type: string
      id:
        type: string
      last_synced_at:
        type: string
      name:
        type: string
      sync_error:
        type: string
      sync_status:
        $ref: '#/definitions/SyncStatus'
    type: object
  KnowledgeIndex:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      driver:
        type: string
      id:
        type: string
      knowledge_id:
        type: string
      name:
        type: string
    type: object
  KnowledgeIndexResponse:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      driver:
        type: string
      id:
        type: string
      knowledge_id:
        type: string
      name:
        type: string
    type: object
  KnowledgeIndexesPage:
    properties:
      data:
        items:
          $ref: '#/definitions/KnowledgeIndex'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
  KnowledgeItem:
    properties:
      attributes:
        $ref: '#/definitions/JSONMap'
      external_id:
        type: string
      hash:
        type: string
      id:
        type: string
      kind:
        $ref: '#/definitions/SourceItemKind'
      knowledge_id:
        type: string
      last_modified:
        type: string
      metadata:
        $ref: '#/definitions/JSONMap'
      name:
        type: string
    type: object
  KnowledgeItemsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/KnowledgeItem'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
  KnowledgeResponse:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      description:
        type: string
      driver:
        type: string
      id:
        type: string
      last_synced_at:
        type: string
      name:
        type: string
      sync_error:
        type: string
      sync_status:
        $ref: '#/definitions/SyncStatus'
    type: object
  KnowledgesPage:
    properties:
      data:
        items:
          $ref: '#/definitions/Knowledge'
        type: array
      skip:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
...
  SourceItemKind:
    enum:
    - text
    - document
    - image
    - audio
    - video
    - structured
    - unknown
    type: string
    x-enum-varnames:
    - SourceItemKindText
    - SourceItemKindDocument
    - SourceItemKindImage
    - SourceItemKindAudio
    - SourceItemKindVideo
    - SourceItemKindStructured
    - SourceItemKindUnknown
  SyncStatus:
    enum:
    - pending
    - in_progress
    - failed
    - success
    type: string
    x-enum-varnames:
    - SyncStatusPending
    - SyncStatusInProgress
    - SyncStatusFailed
    - SyncStatusSuccess
...
  TestConnectionDTO:
    properties:
      configuration:
        $ref: '#/definitions/JSONMap'
      driver:
        maxLength: 255
        type: string
    required:
    - configuration
    - driver
    type: object
...
  UpdateKnowledgeDTO:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        type: string
    type: object
  UpdateKnowledgeIndexDTO:
    properties:
      name:
        maxLength: 255
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Snipet platform.
  title: Snipet API
  version: "1.0"
paths:
...
  /knowledge:
    get:
      description: Lists knowledge bases, with optional pagination.
      parameters:
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KnowledgesPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: List knowledge bases
      tags:
      - knowledge
    post:
      consumes:
      - application/json
      description: Creates a new knowledge base.
      parameters:
      - description: Knowledge data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateKnowledgeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateKnowledgeResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: Create knowledge base
      tags:
      - knowledge
  /knowledge/{id}:
    delete:
      description: Deletes a knowledge base by ID.
      parameters:
      - description: Knowledge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      summary: Delete knowledge base
      tags:
      - knowledge
    get:
      description: Returns a knowledge base by ID.
      parameters:
      - description: Knowledge ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KnowledgeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      summary: Get knowledge base
      tags:
      - knowledge
    put:
      consumes:
      - application/json
      description: Updates a knowledge base by ID.
      parameters:
      - description: Knowledge ID
        in: path
        name: id
        required: true
        type: string
      - description: Knowledge data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateKnowledgeDTO'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: Update knowledge base
      tags:
      - knowledge
  /knowledge/{id}/items:
    get:
      description: Lists items of a knowledge base, with optional pagination.
      parameters:
      - description: Knowledge ID
        in: path
        name: id
        required: true
        type: string
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KnowledgeItemsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: List knowledge items
      tags:
      - knowledge
  /knowledge/{id}/sync:
    post:
      description: Triggers a sync of a knowledge base's items from its source.
      parameters:
      - description: Knowledge ID
        in: path
        name: id
        required: true
        type: string
      - description: Force a full resync
        in: query
        name: force
        type: boolean
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: Sync knowledge base
      tags:
      - knowledge
  /knowledge/{knowledge_id}/index:
    get:
      description: Lists indexes of a knowledge base, with optional pagination.
      parameters:
      - description: Knowledge ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KnowledgeIndexesPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List knowledge indexes
      tags:
      - knowledge-index
    post:
      consumes:
      - application/json
      description: Creates a new index for a knowledge base.
      parameters:
      - description: Knowledge ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: Index data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateKnowledgeIndexDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/KnowledgeIndexResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Create knowledge index
      tags:
      - knowledge-index
  /knowledge/{knowledge_id}/index/{id}:
    delete:
      description: Deletes a knowledge index by ID.
      parameters:
      - description: Knowledge ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: Index ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete knowledge index
      tags:
      - knowledge-index
    get:
      description: Returns a knowledge index by ID.
      parameters:
      - description: Knowledge ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: Index ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KnowledgeIndexResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Get knowledge index
      tags:
      - knowledge-index
    put:
      consumes:
      - application/json
      description: Updates a knowledge index by ID.
      parameters:
      - description: Knowledge ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: Index ID
        in: path
        name: id
        required: true
        type: string
      - description: Index data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateKnowledgeIndexDTO'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Update knowledge index
      tags:
      - knowledge-index
  /knowledge/{knowledge_id}/index/{id}/items:
    get:
      description: Lists items of a knowledge index, with optional pagination.
      parameters:
      - description: Knowledge ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: Index ID
        in: path
        name: id
        required: true
        type: string
      - description: Page size
        in: query
        name: take
        type: integer
      - description: Page offset
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/IndexedKnowledgeItemsPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List indexed knowledge items
      tags:
      - knowledge-index
  /knowledge/drivers:
    get:
      description: Lists the available knowledge source drivers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DriversDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: List knowledge source drivers
      tags:
      - knowledge
  /knowledge/index/drivers:
    get:
      description: Lists the available knowledge index drivers.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DriversDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: List knowledge index drivers
      tags:
      - knowledge-index
  /knowledge/test-connection:
    post:
      consumes:
      - application/json
      description: Tests a knowledge source driver configuration without persisting
        it.
      parameters:
      - description: Connection data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/TestConnectionDTO'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: Test knowledge connection
      tags:
      - knowledge
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"