# internal/module/knowledge-index/dto.go
package knowledgeindex

import (
	"github.com/usesnipet/snipet/internal/filter"
	"github.com/usesnipet/snipet/internal/model"
	"github.com/usesnipet/snipet/internal/page"
	"github.com/usesnipet/snipet/pkg/driver"
	"github.com/usesnipet/snipet/pkg/jsonx"
)

// KnowledgeIndexResponse, KnowledgeIndexesPage and IndexedKnowledgeItemsPage
// exist so swagger annotations in this package can reference them without
// importing internal/model or internal/page directly.
type KnowledgeIndexResponse = model.KnowledgeIndex

type KnowledgeIndexesPage = page.Paginated[model.KnowledgeIndex]

type IndexedKnowledgeItemsPage = page.Paginated[model.IndexedKnowledgeItem]

type CreateKnowledgeIndexDTO struct {
	Name          string        `json:"name" validate:"required,max=255"`
	Driver        string        `json:"driver" validate:"required,max=100"`
	Configuration jsonx.JSONMap `json:"configuration" validate:"required"`
}

type UpdateKnowledgeIndexDTO struct {
	Name *string `json:"name" validate:"omitempty,max=255"`
}

type FilterKnowledgeIndexDTO struct {
	Take *int `form:"take" validate:"omitempty,min=1"`
	Skip *int `form:"skip" validate:"omitempty,min=0"`
}

func (dto *FilterKnowledgeIndexDTO) ToFilter() *filter.Options[model.KnowledgeIndex] {
	return filter.New[model.KnowledgeIndex](
		filter.PtrTake(dto.Take),
		filter.PtrSkip(dto.Skip),
	)
}

type FilterIndexedKnowledgeItemDTO struct {
	Take *int `form:"take" validate:"omitempty,min=1"`
	Skip *int `form:"skip" validate:"omitempty,min=0"`
}

func (dto *FilterIndexedKnowledgeItemDTO) ToFilter() *filter.Options[model.IndexedKnowledgeItem] {
	return filter.New[model.IndexedKnowledgeItem](
		filter.PtrTake(dto.Take),
		filter.PtrSkip(dto.Skip),
	)
}

type DriversDTO struct {
	IndexDrivers []driver.Info `json:"index_drivers"`
}